refactor(media-components): rename files to match the housing component name
